refactor: 画像アップロード定数を lib/image に移して依存方向を修正
lib 層が components を参照していた問題を解消し、paste/upload の想定外例外も catch する。
